Move tests to `test.rs`
