+++ packages/wasm-utils/src/proof/mod.rs
@@ -32,6 +32,7 @@ use crate::{
 
 mod anchor;
 mod mixer;
+mod test;
 #[cfg(test)]
 mod test_utils;
 mod vanchor;
@@ -1379,280 +1380,4 @@ pub fn validate_proof(proof: &Proof, vk: JsString, curve: WasmCurve) -> Result<b
 	.map_err(|e| OperationError::new_with_message(OpStatusCode::InvalidProof, e.to_string()))?;
 
 	Ok(is_valid)
-}
-#[cfg(test)]
-mod test {
-	use ark_std::UniformRand;
-	use arkworks_setups::common::{prove, prove_unchecked, setup_keys, setup_keys_unchecked, verify};
-	use wasm_bindgen_test::*;
-
-	use crate::proof::test_utils::{
-		generate_anchor_test_setup, generate_mixer_test_setup, generate_vanchor_test_rust_setup,
-		generate_vanchor_test_setup, AnchorTestSetup, MixerTestSetup, VAnchorTestSetup, ANCHOR_NOTE_V1_X5_4,
-		ANCHOR_NOTE_V2_X5_4, DECODED_SUBSTRATE_ADDRESS, MIXER_NOTE_V1_X5_5, VANCHOR_NOTE_V2_X5_4,
-	};
-	use crate::{VAnchorR1CSProverBn254_30_2_2_2, DEFAULT_LEAF};
-
-	use super::*;
-
-	const TREE_DEPTH: usize = 30;
-
-	#[wasm_bindgen_test]
-	fn mixer_js_setup() {
-		let MixerTestSetup {
-			relayer,
-			recipient,
-			proof_input_builder,
-			leaf_bytes,
-			..
-		} = generate_mixer_test_setup(DECODED_SUBSTRATE_ADDRESS, DECODED_SUBSTRATE_ADDRESS, MIXER_NOTE_V1_X5_5);
-
-		let truncated_substrate_relayer_address = truncate_and_pad(&relayer);
-		let truncated_substrate_recipient_address = truncate_and_pad(&recipient);
-
-		let proof_input = proof_input_builder.build().unwrap();
-		let mixer_input = proof_input.mixer_input().unwrap();
-
-		assert_eq!(
-			hex::encode(mixer_input.recipient),
-			hex::encode(&truncated_substrate_recipient_address)
-		);
-		assert_eq!(
-			hex::encode(mixer_input.relayer),
-			hex::encode(&truncated_substrate_relayer_address)
-		);
-
-		assert_eq!(mixer_input.refund, 1);
-		assert_eq!(mixer_input.fee, 5);
-
-		assert_eq!(mixer_input.leaf_index, 0);
-		assert_eq!(hex::encode(&mixer_input.leaves[0]), hex::encode(leaf_bytes));
-	}
-
-	#[wasm_bindgen_test]
-	fn anchor_js_setup() {
-		let AnchorTestSetup {
-			relayer,
-			recipient,
-			proof_input_builder,
-			roots_raw,
-			leaf_bytes,
-			..
-		} = generate_anchor_test_setup(
-			DECODED_SUBSTRATE_ADDRESS,
-			DECODED_SUBSTRATE_ADDRESS,
-			ANCHOR_NOTE_V1_X5_4,
-		);
-		let anchor_input = proof_input_builder.build().unwrap().anchor_input().unwrap();
-		let truncated_substrate_relayer_address = truncate_and_pad(&relayer);
-		let truncated_substrate_recipient_address = truncate_and_pad(&recipient);
-		assert_eq!(
-			hex::encode(anchor_input.recipient),
-			hex::encode(&truncated_substrate_recipient_address)
-		);
-		assert_eq!(
-			hex::encode(anchor_input.relayer),
-			hex::encode(&truncated_substrate_relayer_address)
-		);
-
-		assert_eq!(hex::encode(anchor_input.refresh_commitment), hex::encode([0u8; 32]));
-		assert_eq!(hex::encode(&anchor_input.roots[0]), hex::encode(&roots_raw[0]));
-		assert_eq!(anchor_input.roots.len(), roots_raw.len());
-
-		assert_eq!(anchor_input.refund, 1);
-		assert_eq!(anchor_input.fee, 5);
-
-		assert_eq!(anchor_input.leaf_index, 0);
-		assert_eq!(hex::encode(&anchor_input.leaves[0]), hex::encode(leaf_bytes));
-	}
-
-	#[wasm_bindgen_test]
-	fn generate_mixer_proof() {
-		let MixerTestSetup {
-			proof_input_builder,
-			vk,
-			..
-		} = generate_mixer_test_setup(DECODED_SUBSTRATE_ADDRESS, DECODED_SUBSTRATE_ADDRESS, MIXER_NOTE_V1_X5_5);
-
-		let proof_input = proof_input_builder.build_js().unwrap();
-		let proof = generate_proof_js(proof_input).unwrap().proof().unwrap();
-
-		let is_valid_proof = verify_unchecked_raw::<Bn254>(&proof.public_inputs, &vk, &proof.proof).unwrap();
-		assert!(is_valid_proof);
-	}
-
-	#[wasm_bindgen_test]
-	fn anchor_proving_v1() {
-		let AnchorTestSetup {
-			proof_input_builder,
-			vk,
-			..
-		} = generate_anchor_test_setup(
-			DECODED_SUBSTRATE_ADDRESS,
-			DECODED_SUBSTRATE_ADDRESS,
-			ANCHOR_NOTE_V1_X5_4,
-		);
-
-		let proof_input = proof_input_builder.build_js().unwrap();
-		let proof = generate_proof_js(proof_input).unwrap().proof().unwrap();
-
-		let is_valid_proof = verify_unchecked_raw::<Bn254>(&proof.public_inputs, &vk, &proof.proof).unwrap();
-		assert!(is_valid_proof);
-	}
-
-	#[wasm_bindgen_test]
-	fn anchor_proving_v2() {
-		let AnchorTestSetup {
-			proof_input_builder,
-			vk,
-			..
-		} = generate_anchor_test_setup(
-			DECODED_SUBSTRATE_ADDRESS,
-			DECODED_SUBSTRATE_ADDRESS,
-			ANCHOR_NOTE_V2_X5_4,
-		);
-
-		let proof_input = proof_input_builder.build_js().unwrap();
-
-		let proof = generate_proof_js(proof_input.clone()).unwrap().proof().unwrap();
-
-		// UNCOMENT FOR DEBUGGING
-		// match proof_input.inner {
-		// 	ProofInput::Mixer(_) => (),
-		// 	ProofInput::Anchor(anchor_proof_input) => {
-		// 		wasm_bindgen_test::console_log!("exponentiation {:?}",
-		// anchor_proof_input.exponentiation); 		wasm_bindgen_test::console_log!("width
-		// {:?}", anchor_proof_input.width); 		wasm_bindgen_test::console_log!("curve
-		// {:?}", anchor_proof_input.curve); 		wasm_bindgen_test::console_log!("backend
-		// {:?}", anchor_proof_input.backend); 		wasm_bindgen_test::console_log!("secret
-		// {:?}", anchor_proof_input.secret); 		wasm_bindgen_test::console_log!("nullifier
-		// {:?}", anchor_proof_input.nullifier); 		wasm_bindgen_test::console_log!("
-		// recipient {:?}", anchor_proof_input.recipient); 		wasm_bindgen_test::
-		// console_log!("relayer {:?}", anchor_proof_input.relayer); 		wasm_bindgen_test::
-		// console_log!("refund {:?}", anchor_proof_input.refund); 		wasm_bindgen_test::
-		// console_log!("fee {:?}", anchor_proof_input.fee); 		wasm_bindgen_test::
-		// console_log!("chain id {:?}", anchor_proof_input.chain_id);
-		// 		wasm_bindgen_test::console_log!("leaves {:?}", anchor_proof_input.leaves);
-		// 		wasm_bindgen_test::console_log!("leaf index {:?}",
-		// anchor_proof_input.leaf_index); 		wasm_bindgen_test::console_log!("roots {:?}",
-		// anchor_proof_input.roots);
-
-		// 		let params = setup_params(ArkCurve::Bn254, 5, 4);
-		// 		let poseidon = Poseidon::<Bn254Fr>::new(params);
-
-		// 		let chain_id = Bn254Fr::from(anchor_proof_input.chain_id);
-		// 		let secret = Bn254Fr::from_le_bytes_mod_order(&anchor_proof_input.secret);
-		// 		let nullifier =
-		// Bn254Fr::from_le_bytes_mod_order(&anchor_proof_input.nullifier); 		let res =
-		// poseidon.hash(&[chain_id, nullifier, secret]).unwrap(); 		let res_bytes =
-		// res.into_repr().to_bytes_le(); 		let nullifier_hash =
-		// poseidon.hash_two(&nullifier, &nullifier).unwrap(); 		let nullifier_hash_bytes
-		// = nullifier_hash.into_repr().to_bytes_le(); 		wasm_bindgen_test::console_log!("
-		// calc leaf {:?}", res_bytes); 		wasm_bindgen_test::console_log!("nullifier hash
-		// {:?}", nullifier_hash_bytes); 	},
-		// };
-
-		// for p in &proof.public_inputs {
-		// 	wasm_bindgen_test::console_log!("{:?}", p);
-		// }
-
-		let is_valid_proof = verify_unchecked_raw::<Bn254>(&proof.public_inputs, &vk, &proof.proof).unwrap();
-		assert!(is_valid_proof);
-	}
-
-	#[wasm_bindgen_test]
-	fn generate_vanchor_proof_input() {
-		let vanchor_note_str = VANCHOR_NOTE_V2_X5_4;
-		let protocol = JsValue::from("vanchor").into();
-		let mut proof_input_builder = JsProofInputBuilder::new(protocol).unwrap();
-
-		let note = JsNote::deserialize(vanchor_note_str).unwrap();
-		let leaf = note.get_leaf_commitment().unwrap();
-		let leaves: Array = vec![leaf].into_iter().collect();
-
-		let mut tree_leaves = LeavesMapInput::new();
-		tree_leaves
-			.set_chain_leaves(2, Leaves::from(JsValue::from(leaves.clone())))
-			.unwrap();
-		tree_leaves
-			.set_chain_leaves(3, Leaves::from(JsValue::from(leaves.clone())))
-			.unwrap();
-
-		let indices: Array = vec![JsValue::from("0"), JsValue::from("0")].into_iter().collect();
-		let roots: Array = vec![
-			Uint8Array::from([0u8; 32].to_vec().as_slice()),
-			Uint8Array::from([0u8; 32].to_vec().as_slice()),
-		]
-		.into_iter()
-		.collect();
-		proof_input_builder.set_leaves_map(tree_leaves).unwrap();
-		proof_input_builder.set_metadata_from_note(&note).unwrap();
-
-		proof_input_builder
-			.set_indices(Indices::from(JsValue::from(indices)))
-			.unwrap();
-		proof_input_builder
-			.set_roots(Leaves::from(JsValue::from(roots)))
-			.unwrap();
-
-		proof_input_builder.set_pk(JsString::from("0000")).unwrap();
-		proof_input_builder.public_amount(JsString::from("3")).unwrap();
-		proof_input_builder.chain_id(JsString::from("4")).unwrap();
-		proof_input_builder.set_ext_data_hash(JsString::from("1111")).unwrap();
-		proof_input_builder
-			.set_vanchor_output_config(
-				OutputUtxoConfig {
-					index: None,
-					amount: 0,
-					chain_id: 3,
-				},
-				OutputUtxoConfig {
-					index: None,
-					amount: 0,
-					chain_id: 3,
-				},
-			)
-			.unwrap();
-		let notes: Array = vec![JsValue::from(note.clone())].into_iter().collect();
-		// TODO Fix  `AsRef<wasm_bindgen::JsValue>` is not implemented for
-		// `note::JsNote`
-		proof_input_builder.set_notes(notes).unwrap();
-		let proof_builder = proof_input_builder.build_js().unwrap();
-		let vanchor_proof_input_payload = proof_builder.inner.vanchor_input().unwrap();
-		assert_eq!(vanchor_proof_input_payload.public_amount, 3);
-		assert_eq!(vanchor_proof_input_payload.chain_id, 4);
-		assert_eq!(vanchor_proof_input_payload.indices, [0u64, 0u64].to_vec());
-		assert_eq!(
-			hex::encode(&vanchor_proof_input_payload.roots[0]),
-			hex::encode([0u8; 32].to_vec())
-		);
-		assert_eq!(vanchor_proof_input_payload.roots.len(), 2);
-		assert_eq!(
-			vanchor_proof_input_payload.backend.to_string(),
-			note.backend.unwrap().to_string()
-		);
-		assert_eq!(vanchor_proof_input_payload.exponentiation, note.exponentiation.unwrap());
-		assert_eq!(vanchor_proof_input_payload.width, note.width.unwrap());
-		assert_eq!(
-			vanchor_proof_input_payload.curve.to_string(),
-			note.curve.unwrap().to_string()
-		);
-		assert_eq!(hex::encode(vanchor_proof_input_payload.pk), "0000");
-	}
-
-	#[wasm_bindgen_test]
-	fn generate_vanchor_proof() {
-		let VAnchorTestSetup {
-			proof_input_builder,
-			roots_raw,
-			notes,
-			leaf_index,
-			vk,
-		} = generate_vanchor_test_rust_setup(DECODED_SUBSTRATE_ADDRESS, DECODED_SUBSTRATE_ADDRESS);
-		let proof_input = proof_input_builder.build_js().unwrap();
-		let proof = generate_proof_js(proof_input).unwrap().vanchor_proof().unwrap();
-		let is_valid_proof = verify_unchecked_raw::<Bn254>(&proof.public_inputs, &vk, &proof.proof).unwrap();
-
-		assert!(is_valid_proof);
-	}
 }
